HDDS-4829. Rename MiniOzoneHACluster to MiniOzoneOMHACluster.
